Refactor(sys): 收窄菜单树持久化契约

+++ sandwish-biz/src/main/java/com/github/thundax/modules/sys/service/impl/MenuServiceImpl.java
@@ -3,7 +3,8 @@ package com.github.thundax.modules.sys.service.impl;
 import com.github.thundax.common.Constants;
 import com.github.thundax.common.collect.ListUtils;
 import com.github.thundax.common.collect.MapUtils;
-import com.github.thundax.common.service.impl.TreeServiceImpl;
+import com.github.thundax.common.persistence.TreeEntity;
+import com.github.thundax.common.service.impl.CrudServiceImpl;
 import com.github.thundax.common.utils.SpringContextHolder;
 import com.github.thundax.common.utils.StringUtils;
 import com.github.thundax.common.utils.redis.RedisClient;
@@ -23,7 +24,7 @@ import java.util.Objects;
  */
 @Service
 @Transactional(readOnly = true)
-public class MenuServiceImpl extends TreeServiceImpl<MenuDao, Menu> implements MenuService {
+public class MenuServiceImpl extends CrudServiceImpl<MenuDao, Menu> implements MenuService {
 
     private String lastCacheVersion;
     private final SignService signService;
@@ -96,8 +97,53 @@ public class MenuServiceImpl extends TreeServiceImpl<MenuDao, Menu> implements M
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void save(Menu menu) {
-        super.save(menu);
+        if (menu.getIsNewRecord()) {
+            menu.preInsert();
 
+            Integer newPosition;
+            if (StringUtils.isNotBlank(menu.getParentId())
+                    && !StringUtils.equals(menu.getParentId(), TreeEntity.ROOT_ID)) {
+                Menu parent = dao.getTreeNode(menu.getParentId());
+                newPosition = parent.getRgt();
+            } else {
+                menu.setParentId(null);
+                Integer maxRgt = dao.getMaxPosition();
+                if (maxRgt == null) {
+                    maxRgt = 0;
+                }
+                newPosition = maxRgt + 1;
+            }
+            menu.setLft(newPosition);
+            menu.setRgt(newPosition + 1);
+            dao.moveTreeRgts(newPosition, 2);
+            dao.moveTreeLfts(newPosition, 2);
+            dao.insert(menu);
+
+        } else {
+            menu.preUpdate();
+            Menu oldBean = dao.getTreeNode(menu.getId());
+            if (!StringUtils.equals(oldBean.getParentId(), menu.getParentId())) {
+                Integer newPosition;
+                if (StringUtils.isNotBlank(menu.getParentId())) {
+                    Menu parent = dao.getTreeNode(menu.getParentId());
+                    newPosition = parent.getRgt();
+                } else {
+                    newPosition = dao.getMaxPosition() + 1;
+                }
+                dao.moveTreeRgts(newPosition, oldBean.treeSpan());
+                dao.moveTreeLfts(newPosition, oldBean.treeSpan());
+
+                oldBean = dao.getTreeNode(menu.getId());
+                int offset = newPosition - oldBean.getLft();
+                dao.moveTreeNodes(oldBean.getLft(), oldBean.getRgt(), offset);
+
+                dao.moveTreeRgts(oldBean.getLft(), -oldBean.treeSpan());
+                dao.moveTreeLfts(oldBean.getLft(), -oldBean.treeSpan());
+            }
+            dao.update(menu);
+        }
+
+        removeAllCache();
         signService.sign(menu.getSignName(), menu.getSignId(), menu.getSignBody());
     }
 
@@ -124,10 +170,67 @@ public class MenuServiceImpl extends TreeServiceImpl<MenuDao, Menu> implements M
     @Transactional(rollbackFor = Exception.class)
     public int delete(Menu menu) {
         dao.deleteMenuRole(menu);
-        int retVal = super.delete(menu);
+        Menu bean = this.get(menu.getId());
+        if (bean == null) {
+            return 0;
+        }
+        dao.moveTreeRgts(bean.getLft(), -bean.treeSpan());
+        dao.moveTreeLfts(bean.getLft(), -bean.treeSpan());
+
+        int retVal = dao.delete(bean);
+
+        removeAllCache();
         signService.deleteSign(menu.getSignName(), menu.getSignId());
 
         return retVal;
+    }
+
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    public void moveTreeNode(Menu from, Menu to, MoveTreeNodeType moveType) {
+        Menu fromBean = dao.getTreeNode(from.getId());
+        Menu toBean = dao.getTreeNode(to.getId());
+
+        int newPosition;
+        String newParentId;
+        if (moveType == MoveTreeNodeType.AFTER) {
+            newPosition = toBean.getRgt() + 1;
+            newParentId = toBean.getParentId();
+
+        } else if (moveType == MoveTreeNodeType.BEFORE) {
+            newPosition = toBean.getLft();
+            newParentId = toBean.getParentId();
+
+        } else if (moveType == MoveTreeNodeType.INSIDE) {
+            newPosition = toBean.getLft() + 1;
+            newParentId = to.getId();
+
+        } else {
+            newPosition = toBean.getRgt();
+            newParentId = to.getId();
+        }
+
+        dao.moveTreeLfts(newPosition, fromBean.treeSpan());
+        dao.moveTreeRgts(newPosition, fromBean.treeSpan());
+
+        fromBean = dao.getTreeNode(from.getId());
+        int offset = newPosition - fromBean.getLft();
+        dao.moveTreeNodes(fromBean.getLft(), fromBean.getRgt(), offset);
+
+        dao.moveTreeLfts(fromBean.getLft(), -fromBean.treeSpan());
+        dao.moveTreeRgts(fromBean.getLft(), -fromBean.treeSpan());
+
+        from.setParentId(newParentId);
+        dao.updateParent(from);
+
+        removeAllCache();
+    }
+
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    public void updateLftRgt(Menu entity) {
+        dao.updateLftRgt(entity);
+        removeAllCache();
     }
 
     @Override
